Arbitrage package moved into trade directory and documentation added to the arbitrage package
